+++ uniswap-v2/logics/impls/router/data.rs
@@ -1,0 +1,11 @@
+use openbrush::traits::{
+    AccountId,
+};
+
+pub const STORAGE_KEY: u32 = openbrush::storage_unique_key!(Data);
+
+#[derive(Default, Debug)]
+#[openbrush::upgradeable_storage(STORAGE_KEY)]
+pub struct Data {
+    pub factory: AccountId,
+}
